@@ -20,14 +20,14 @@ import junit.framework.TestSuite;
 
 import org.junit.internal.builders.AllDefaultPossibilitiesBuilder;
 
-import android.support.test.filters.LargeTest;
-import android.support.test.runner.AndroidJUnitRunner;
+import androidx.test.filters.LargeTest;
+import androidx.test.runner.AndroidJUnitRunner;
 import android.test.ActivityInstrumentationTestCase2;
 
 import static com.example.android.testing.androidjunitrunnersample.HintMatcher.withHint;
-import static android.support.test.espresso.Espresso.onView;
-import static android.support.test.espresso.assertion.ViewAssertions.matches;
-import static android.support.test.espresso.matcher.ViewMatchers.withId;
+import static androidx.test.espresso.Espresso.onView;
+import static androidx.test.espresso.assertion.ViewAssertions.matches;
+import static androidx.test.espresso.matcher.ViewMatchers.withId;
 import static org.hamcrest.CoreMatchers.notNullValue;
 import static org.junit.Assert.assertThat;
 
