@@ -16,7 +16,7 @@
 
 package com.example.android.testing.espresso.intents.AdvancedSample;
 
-import android.support.test.espresso.matcher.BoundedMatcher;
+import androidx.test.espresso.matcher.BoundedMatcher;
 import android.view.View;
 import android.widget.ImageView;
 
