@@ -17,7 +17,7 @@
 package com.example.android.testing.espresso.RecyclerViewSample;
 
 import android.content.Context;
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
