@@ -19,8 +19,8 @@ package com.example.android.testing.espresso.web.BasicSample;
 import android.app.Activity;
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.VisibleForTesting;
+import androidx.annotation.NonNull;
+import androidx.annotation.VisibleForTesting;
 import android.text.TextUtils;
 import android.webkit.WebView;
 import android.webkit.WebViewClient;
